+++ packages/pigeon/e2e_tests/test_objc/ios/Runner/MyNestedApi.m
@@ -5,7 +5,7 @@
 #import "MyNestedApi.h"
 
 @implementation MyNestedApi
-- (ACSearchReply *)search:(ACNested *)input error:(FlutterError **)error {
+- (ACSearchReply *)searchNested:(ACNested *)input error:(FlutterError **)error {
   ACSearchReply *reply = [[ACSearchReply alloc] init];
   reply.result = [NSString stringWithFormat:@"Hello %@!", input.request.query];
   reply.state = ACRequestStateSuccess;