+++ packages/pigeon/platform_tests/ios_unit_tests/ios/RunnerTests/EchoMessenger.h
@@ -7,7 +7,10 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+/// A FlutterBinaryMessenger who replies with the first argument sent to it.
 @interface EchoBinaryMessenger : NSObject<FlutterBinaryMessenger>
+- (instancetype)init NS_UNAVAILABLE;
+- (instancetype)initWithCodec:(NSObject<FlutterMessageCodec>*)codec;
 @end
 
 NS_ASSUME_NONNULL_END